Rename filescope to loadfile
This makes it consistent with the option (--dist=loadfile) and the other
scheduling file (loadscope.py)
